Adjust Index + selected menu item
