+++ src/components/header.tsx
@@ -6,14 +6,15 @@ import { graphql, useStaticQuery } from 'gatsby';
 import { LocalizedLink } from 'gatsby-theme-i18n';
 import ThemeContext from '../context/ThemeContext';
 
+// Helpers
 const styleToggle = {
   fontWeight: 'bold',
 };
 const defaultStyle = {};
+const isHome = (path: string) => ['/', '/fr', '/fr/'].includes(path);
 
 const SelectLanguage: React.FC<any> = (props) => {
   const { langs, pathname, locale } = props;
-  const isHome = (path: string) => ['/', '/fr', '/fr/'].includes(path);
 
   const links = langs.map(({ langKey }: any) => {
     let to;
@@ -43,16 +44,23 @@ const SelectLanguage: React.FC<any> = (props) => {
   );
 };
 
-const getMenuItems = (menu: any, pathname: string, locale: string) => {
-  const delocalizedSlug = locale === 'en' ? pathname : pathname.slice(3);
-  return menu.map((item: any) => (
-    <LocalizedLink to={item.slug} key={item.slug}>
-      <li style={item.slug === delocalizedSlug ? styleToggle : defaultStyle}>
-        {item.label}
-      </li>
-    </LocalizedLink>
-  ));
-};
+const getMenuItems = (menu: any, pathname: string, locale: string) =>
+  menu.map((item: any) => {
+    const delocalizedPath = locale === 'en' ? pathname : pathname.slice(3);
+    let linkStyle;
+    if (isHome(pathname) && isHome(item.slug)) {
+      linkStyle = styleToggle;
+    } else if (!isHome(item.slug) && delocalizedPath.startsWith(item.slug)) {
+      linkStyle = styleToggle;
+    } else {
+      linkStyle = defaultStyle;
+    }
+    return (
+      <LocalizedLink to={item.slug} key={item.slug}>
+        <li style={linkStyle}>{item.label}</li>
+      </LocalizedLink>
+    );
+  });
 
 const Header: React.FC<any> = (props) => {
   const {
